Adapted build
