+++ labs/unicorn-store/infrastructure/cdk/src/main/java/com/unicorn/UnicornStoreStack.java
@@ -3,16 +3,15 @@ package com.unicorn;
 import com.unicorn.constructs.UnicornStoreBasic;
 import com.unicorn.constructs.UnicornStoreMicronaut;
 import com.unicorn.constructs.UnicornStoreSpringNative;
+import com.unicorn.core.InfrastructureStack;
 import software.amazon.awscdk.*;
 import software.amazon.awscdk.services.apigateway.LambdaRestApi;
 import software.amazon.awscdk.services.apigateway.RestApi;
 import software.amazon.awscdk.services.lambda.Code;
 import software.amazon.awscdk.services.lambda.Function;
 import software.amazon.awscdk.services.lambda.Runtime;
-import software.amazon.awscdk.services.logs.RetentionDays;
 import software.constructs.Construct;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
