+++ labs/unicorn-store/infrastructure/cdk/src/main/java/com/unicorn/constructs/UnicornStoreBasic.java
@@ -1,6 +1,6 @@
 package com.unicorn.constructs;
 
-import com.unicorn.InfrastructureStack;
+import com.unicorn.core.InfrastructureStack;
 import software.amazon.awscdk.CfnOutput;
 import software.amazon.awscdk.CfnOutputProps;
 import software.amazon.awscdk.Duration;
@@ -11,7 +11,6 @@ import software.amazon.awscdk.services.apigateway.RestApi;
 import software.amazon.awscdk.services.lambda.Code;
 import software.amazon.awscdk.services.lambda.Function;
 import software.amazon.awscdk.services.lambda.Runtime;
-import software.amazon.awscdk.services.logs.RetentionDays;
 import software.constructs.Construct;
 
 import java.util.HashMap;
