+++ labs/unicorn-store/infrastructure/cdk/src/main/java/com/unicorn/constructs/UnicornStoreMicronaut.java
@@ -1,13 +1,12 @@
 package com.unicorn.constructs;
 
-import com.unicorn.InfrastructureStack;
+import com.unicorn.core.InfrastructureStack;
 import software.amazon.awscdk.*;
 import software.amazon.awscdk.services.apigateway.LambdaRestApi;
 import software.amazon.awscdk.services.apigateway.RestApi;
 import software.amazon.awscdk.services.lambda.Code;
 import software.amazon.awscdk.services.lambda.Function;
 import software.amazon.awscdk.services.lambda.Runtime;
-import software.amazon.awscdk.services.logs.RetentionDays;
 import software.constructs.Construct;
 
 import java.util.HashMap;
