+++ labs/unicorn-store/infrastructure/cdk/src/main/java/com/unicorn/constructs/UnicornStoreSpringNative.java
@@ -1,10 +1,9 @@
 package com.unicorn.constructs;
 
-import com.unicorn.InfrastructureStack;
+import com.unicorn.core.InfrastructureStack;
 import software.amazon.awscdk.CfnOutput;
 import software.amazon.awscdk.CfnOutputProps;
 import software.amazon.awscdk.Duration;
-import software.amazon.awscdk.services.apigateway.LambdaIntegration;
 import software.amazon.awscdk.services.apigateway.LambdaRestApi;
 import software.amazon.awscdk.services.apigateway.RestApi;
 import software.amazon.awscdk.services.lambda.Code;
@@ -12,7 +11,6 @@ import software.amazon.awscdk.services.lambda.Function;
 import software.amazon.awscdk.services.lambda.Runtime;
 import software.constructs.Construct;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
